requirements.txt
Flask
marshmallow-sqlalchemy
flask-marshmallow
Flask-Migrate
Flask-Script
Flask-JWT-Extended
